LayoutTests/svg/animations/svgenum-animation-4-expected.txt

SVG 1.1 dynamic animation tests

Test CompositeOperationType enumeration animations

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS overComposite1.operator.animVal is SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_OVER
PASS overComposite1.operator.baseVal is SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_OVER
PASS overComposite1.operator.animVal is SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_IN
PASS overComposite1.operator.baseVal is SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_OVER
PASS overComposite1.operator.animVal is SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_IN
PASS overComposite1.operator.baseVal is SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_OVER
PASS overComposite1.operator.animVal is SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_OUT
PASS overComposite1.operator.baseVal is SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_OVER
PASS overComposite1.operator.animVal is SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_OUT
PASS overComposite1.operator.baseVal is SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_OVER
PASS overComposite1.operator.animVal is SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_ATOP
PASS overComposite1.operator.baseVal is SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_OVER
PASS overComposite1.operator.animVal is SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_ATOP
PASS overComposite1.operator.baseVal is SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_OVER
PASS overComposite1.operator.animVal is SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_XOR
PASS overComposite1.operator.baseVal is SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_OVER
PASS overComposite1.operator.animVal is SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_XOR
PASS overComposite1.operator.baseVal is SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_OVER
PASS overComposite1.operator.animVal is SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_ARITHMETIC
PASS overComposite1.operator.baseVal is SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_OVER
PASS overComposite1.operator.animVal is SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_ARITHMETIC
PASS overComposite1.operator.baseVal is SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_OVER
PASS overComposite1.operator.animVal is SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_OVER
PASS overComposite1.operator.baseVal is SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_OVER
PASS successfullyParsed is true

TEST COMPLETE
